+++ R/lifeduration.R
@@ -6,6 +6,7 @@
 #' @param at Date old version of package. Default: NULL
 #' @param lib.loc character vector. Is omitted for non NULL version. Default: `.libPaths()`
 #' @param repos character vector base URLs of the repositories to use. By default checking CRAN and newest Bioconductor. Default `pacs::biocran_repos()`
+#' @param source character one of `c("metadb", "cran")`. Using the `MEATCRAN` DB or the direct web page download from CRAN. Default: `"metadb"`
 #' @return `difftime`, number of days package version was the newest one.
 #' @note Function will scrap two github CRAN mirror and CRAN URL. Works mainly with CRAN packages.
 #' Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
@@ -23,11 +24,14 @@ pac_lifeduration <- function(pac,
                              version = NULL,
                              at = NULL,
                              lib.loc = .libPaths(),
-                             repos = biocran_repos()) {
+                             repos = biocran_repos(),
+                             source = c("metadb", "cran")) {
   stopifnot(length(pac) == 1 && is.character(pac))
   stopifnot(!all(c(!is.null(version), !is.null(at))))
   stopifnot(is.null(lib.loc) || (all(lib.loc %in% .libPaths()) && (length(list.files(lib.loc)) > 0)))
   stopifnot(is.null(version) || (length(version) == 1 && is.character(version)))
+
+  source <- match.arg(source)
 
   ison_cran <- is_isin(pac, "https://cran.rstudio.com/")
   last_version <- pac_last(pac, repos = repos)
@@ -88,6 +92,7 @@ pac_lifeduration <- function(pac,
 #' @param flavors character vector of CRAN machines to consider, which might be retrieved with `pacs::cran_flavors()$Flavor`. By default all CRAN machines are considered, NULL value. Default NULL
 #' @param lib.loc character vector. Is omitted for non NULL version. Default: `.libPaths()`
 #' @param repos character the base CRAN URL of the repository to use. Default "https://cran.rstudio.org"
+#' @param source character one of `c("metadb", "cran")`. Using the `MEATCRAN` DB or the direct web page download from CRAN. Default: `"metadb"`
 #' @return logical if a package is healthy.
 #' @note Function will scrap two/tree CRAN URLS. Works only with CRAN packages.
 #' The newest release are checked for warnings/errors on R CRAN check page.
@@ -108,11 +113,14 @@ pac_health <- function(pac,
                        scope = c("ERROR", "FAIL"),
                        flavors = NULL,
                        lib.loc = .libPaths(),
-                       repos = "https://cran.rstudio.com/") {
+                       repos = "https://cran.rstudio.com/",
+                       source = c("metadb", "cran")) {
   stopifnot(length(pac) == 1 && is.character(pac))
   stopifnot(!all(c(!is.null(version), !is.null(at))))
   stopifnot(is.null(version) || (length(version) == 1 && is.character(version)))
   stopifnot(length(repos) == 1 && is.character(repos))
+
+  source <- match.arg(source)
 
   if (pac %in% pacs_base()) {
     return(TRUE)